+++ hooks/useAnsweredPosts.ts
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from '@tanstack/react-query'
-import { fetchSlicePosts } from './fetchSlicePosts'
+import { fetchSlicePosts } from '../app/client/fetchSlicePosts'
 
 export const useAnsweredPosts = () =>
   useInfiniteQuery({
