+++ hooks/useDeletePost.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query'
-import { deletePost } from './deletePost'
+import { deletePost } from '../app/client/deletePost'
 
 export const useDeletePost = () =>
   useMutation(async (contentId: string) => await deletePost(contentId))
