+++ hooks/useDeleteReply.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query'
-import { deleteReply } from './deleteReply'
+import { deleteReply } from '../app/client/deleteReply'
 
 export const useDeleteReply = () =>
   useMutation(async (contentId: string) => await deleteReply(contentId))
